.history/requirements_20240328114219.txt
numpy==1.19.5
matplotlib==3.4.2
# torchvision==0.8.0
bcolz
tqdm==4.62.0
opencv_python==4.5.3.56
# torch==1.7.0
scipy==1.6.2
Pillow==8.3.2
scikit_learn==0.24.2
tensorboardX==2.4
